+++ tests/_cvodes_cxx.pyx
@@ -6,6 +6,7 @@ cdef extern from "testing_utils.hpp":
     cppclass Decay:
         Decay(double)
 
+
 cdef class PyDecay:
     cdef Decay *thisptr
 
@@ -18,4 +19,5 @@ cdef class PyDecay:
     def adaptive(self, double y0, double t):
         cdef:
             vector[int] root_indices
-        return simple_adaptive[Decay](self.thisptr, [1e-10], 1e-10, 1, &y0, 0.0, t, root_indices)
+            int ADAMS = 1  # grep "define CV_ADAMS" cvodes.h
+        return simple_adaptive[Decay](self.thisptr, [1e-10], 1e-10, ADAMS, &y0, 0.0, t, root_indices)