+++ tests/test_cvodes_cxx.cpp
@@ -7,7 +7,7 @@
 using SVector = sundials_cxx::nvector_serial::Vector;
 
 
-TEST_CASE( "methods" "[CVodeIntegrator]" ) {
+TEST_CASE( "methods", "[CVodeIntegrator]" ) {
     auto intgr = cvodes_cxx::CVodeIntegrator(cvodes_cxx::LMM::Adams, cvodes_cxx::IterType::Functional);
     std::vector<double> y(1, 1.0);
     double t, yref;
@@ -22,7 +22,7 @@ TEST_CASE( "methods" "[CVodeIntegrator]" ) {
 }
 
 
-TEST_CASE( "decay_adaptive" "[simple_adaptive]" ) {
+TEST_CASE( "decay_adaptive", "[simple_adaptive]" ) {
     Decay odesys(1.0);
     double y0 = 1.0;
     std::vector<int> root_indices;
@@ -34,4 +34,20 @@ TEST_CASE( "decay_adaptive" "[simple_adaptive]" ) {
         REQUIRE( std::abs(std::exp(-tout[i]) - yout[i]) < 1e-8 );
     }
     REQUIRE( odesys.last_integration_info["n_steps"] > 1 );
+    REQUIRE( odesys.last_integration_info["n_steps"] < 997 );
+}
+
+
+TEST_CASE( "decay_adaptive_dx_max", "[simple_adaptive]" ) {
+    Decay odesys(1.0);
+    double y0 = 1.0;
+    std::vector<int> root_indices;
+    auto tout_yout = cvodes_cxx::simple_adaptive(&odesys, {1e-10}, 1e-10, CV_ADAMS, &y0, 0.0, 1.0, root_indices, 0.0, 0.0, 1e-3, 1100);
+    auto& tout = tout_yout.first;
+    auto& yout = tout_yout.second;
+    REQUIRE( tout.size() == yout.size() );
+    for (uint i = 0; i < tout.size(); ++i){
+        REQUIRE( std::abs(std::exp(-tout[i]) - yout[i]) < 1e-8 );
+    }
+    REQUIRE( odesys.last_integration_info["n_steps"] > 998 );
 }